requirements.txt
flask==2.3.3
line-bot-sdk==3.5.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
google-generativeai==0.3.2
python-dotenv==1.0.0
lxml==4.9.3
